+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 - pytest integration bug fixed
 - ViewMethod copy bug fixed
 - pydantic required version increased
+- openapi/openrpc specification definitions support implemented
 
 
 1.3.0 (2021-08-13)